requirements.txt
alembic==0.9.4
amqp==2.2.2
aniso8601==3.0.0
appdirs==1.4.3
appnope==0.1.0
APScheduler==3.5.1
asn1crypto==0.24.0
astroid==1.6.4
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.3
blinker==1.4
bottle==0.12.13
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.2.2
decorator==4.0.11
dominate==2.3.1
enum34==1.1.6
ez-setup==0.9
Flask==0.12.2
Flask-APScheduler==1.10.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.0
Flask-Moment==0.6.0
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.5
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.44
packaging==16.8
pathlib2==2.3.0
pbr==3.0.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.4.2
pylint==2.0.0.dev0
pymongo==3.6.1
PyMySQL==0.9.1
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
pyzabbix==0.7.4
qiniu==7.1.4
redis==2.10.5
requests==2.6.0
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.10
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.22
vine==1.1.4
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
wget==3.2
wrapt==1.10.11
WTForms==2.1
WTForms-JSON==0.3.3
xmltodict==0.9.2
